requirements.txt
python-fasthtml
fh_frankenui
fh_matplotlib
scipy